requirements_windows.txt
ultralytics
opencv-python
pillow
numpy
matplotlib
torch
torchvision
segmentation-models-pytorch
timm
